requirements.txt
django==5.2.5
gunicorn==23.0.0
python-decouple==3.8
pillow==11.3.0
django-crum==0.7.9
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
celery==5.5.3
redis==6.4.0
django-celery-beat==2.8.1
flower==2.0.1
django-redis==6.0.0
pylint==3.3.8
pylint-django==2.6.1
django-ckeditor==6.7.3
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
